General GUI
Got the General components of the GUI done. The lists don't work yet since we need to grab that info from the server, I think. Also added Strings for the getText elements to make it easier to get things going.

Will rename the variables later (paneCatalog is still there in admin)