A simple GUI test with possible workaround
Testing the updating lists. Works but unfortunately, it only works because of the mimicker arraylist that tries to mimic actions that are going on server side (note: did not implement any of the exit app stuff so exiting out will crash the server).

Basically this test shows I need to do a LOT of cleaning in the Admin Controller. No need for all that stuff to update a list xD
We can delete this once we're ready to turn in the project