Organized Controller + Fixed Order Report
Controller is organized a bit more.
Order list is fixed. Appearently it didn't like the output.reset for some reason. I think it's fine for this method because Admins cannot add to the Order Report, only clients.
Also added warranty setters and getters to Electronic

Once we get all the lists going, I'm going to fix a lot of the GUI components and make it look better. Right now all the widths/heights are all out of wack

Almost there!