Catalog App up and running (except for remove)
I'm still working on figuring out how to fix remove. It doesn't even show an update on the server side so it's possible my implementation of it is bad in Order class (I used a collection) Could also be that it got messed up when I was changing things in Company. For now, Feel free to go ahead and get the same closing app safety nets as in adminApp :)